Bank accounts management

+++ app/helpers/bank_accounts_helper.rb
@@ -1,5 +1,7 @@
 module BankAccountsHelper
   def bank_account_delete(bank_account)
-    link_to(image_tag("delete.png", :alt => t(".delete"), :title => t(".delete")), user_bank_account_path(bank_account), :method => :delete, :confirm => t(".confirm"))
+    if bank_account.wire_transfers.blank? && (bank_account.state != 'verified')
+      link_to(image_tag("delete.png", :alt => t(".delete"), :title => t(".delete")), user_bank_account_path(bank_account), :method => :delete, :confirm => t(".confirm"))
+    end
   end
 end
